+++ src/main/java/Courier.java
@@ -1,40 +1,12 @@
-public class Courier {
+import lombok.Data;
 
+@Data
+public class Courier {
     private String login;
     private String password;
-    private String firstName;
 
-
-    public Courier(){
-
-    }
-
-    public Courier(String login, String password){
+    public Courier(String login, String password) {
         this.login = login;
         this.password = password;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
     }
 }
