+++ src/main/java/Order.java
@@ -1,5 +1,9 @@
 import java.util.List;
 
+
+import lombok.Data;
+
+@Data
 public class Order {
     private String firstName;
     private String lastName;
@@ -20,82 +24,6 @@ public class Order {
         this.rentTime = rentTime;
         this.deliveryDate = deliveryDate;
         this.comment = comment;
-        this.color = color;
-    }
-
-    public Order(){
-
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getMetroStation() {
-        return metroStation;
-    }
-
-    public void setMetroStation(String metroStation) {
-        this.metroStation = metroStation;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public int getRentTime() {
-        return rentTime;
-    }
-
-    public void setRentTime(int rentTime) {
-        this.rentTime = rentTime;
-    }
-
-    public String getDeliveryDate() {
-        return deliveryDate;
-    }
-
-    public void setDeliveryDate(String deliveryDate) {
-        this.deliveryDate = deliveryDate;
-    }
-
-    public String getComment() {
-        return comment;
-    }
-
-    public void setComment(String comment) {
-        this.comment = comment;
-    }
-
-    public List<String> getColor() {
-        return color;
-    }
-
-    public void setColor(List<String> color) {
         this.color = color;
     }
 }
