+++ src/main/java/ProjectApiClient.java
@@ -1,16 +1,12 @@
 import io.restassured.response.Response;
 
-import java.util.HashMap;
-import java.util.Map;
-
-public class CourierApiClient extends BaseHttpClient {
+public class ProjectApiClient extends BaseHttpClient {
     private final String COURIER_CREATE_REQUEST = "api/v1/courier";
     private final String COURIER_LOGIN_REQUEST = "api/v1/courier/login";
     private final String DELETE_REQUEST = "api/v1/courier/";
     private final String CREATE_ORDER_REQUEST = "api/v1/orders/";
     private final String GET_ORDER_BY_TRACK = "api/v1/orders/track?t=";
     private final String GET_ORDERS_LIST = "/api/v1/orders/";
-
     private final String CANCEL_ORDER = "api/v1/orders/cancel?track=";
