+++ src/test/java/CourierCreatingTest.java
@@ -4,6 +4,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import static org.apache.http.HttpStatus.*;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
@@ -11,10 +12,10 @@ import static org.hamcrest.CoreMatchers.notNullValue;
 public class CourierCreatingTest {
     String login = "trevorfillipscorp";
     String password = "1234";
-    Steps step;
+    CourierSteps step;
     @Before
     public void setUp(){
-      step = new Steps();
+      step = new CourierSteps();
     } //Создаём объект класса с шагами
 
     @Test
@@ -23,14 +24,14 @@ public class CourierCreatingTest {
     public void courierCreatingTest(){
        step.courierCreating(login, password) //Создаём курьера
                .then()
-               .statusCode(201) //Проверяем статус-код
+               .statusCode(SC_CREATED) //Проверяем статус-код
                .and()
                .assertThat().body("ok", equalTo(true)); //Проверяем тело ответа
 
        step.courierLogin(login, password) //Проверка того, что курьер создан - авторизация
                .then().assertThat()
                .body("id", notNullValue()) //Ответ с телом id курьера
-               .and().statusCode(200); //Статус-код авторизации
+               .and().statusCode(SC_OK); //Статус-код авторизации
     }
 
     @Test
@@ -40,7 +41,7 @@ public class CourierCreatingTest {
         step.courierCreating(login, password); //Создаём курьера
         step.courierCreating(login, password) //Пытаемся повторно создать курьера
                 .then()
-                .statusCode(409) //Статус-код о существуюшем курьере
+                .statusCode(SC_CONFLICT) //Статус-код о существуюшем курьере
                 .and()
                 .body("message", equalTo("Этот логин уже используется. Попробуйте другой.")); //Тело ответа
     }
@@ -51,7 +52,7 @@ public class CourierCreatingTest {
     public void testCreatingCourierWithMissingFields(){
         step.courierCreating(null, password)
                 .then()
-                .statusCode(400)
+                .statusCode(SC_BAD_REQUEST)
                 .and()
                 .body("message", equalTo("Недостаточно данных для создания учетной записи")); //Создаём курьера с null вместо логина
     }
