+++ src/test/java/GetOrdersReturnsListOfOrdersTest.java
@@ -1,11 +1,11 @@
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.util.List;
 
+import static org.apache.http.HttpStatus.*;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
 public class GetOrdersReturnsListOfOrdersTest {
@@ -18,11 +18,11 @@ public class GetOrdersReturnsListOfOrdersTest {
     private String deliveryDate = "2024-02-15";
     private String comment = "Come to Vinewood Hills";
     private List<String> color;
-    Steps step;
+    OrderSteps step;
 
     @Before
     public void setUp(){
-        step = new Steps(); //Создаём объект класса с шагами
+        step = new OrderSteps(); //Создаём объект класса с шагами
     }
 
     @Test
@@ -31,7 +31,7 @@ public class GetOrdersReturnsListOfOrdersTest {
     public void testGetOrderReturnListOfOrders(){
         step.getOrdersList() //Выполняем гет-запрос на получение списка заказов
                 .then()
-                .statusCode(200) //Ожидание статус-кода 200
+                .statusCode(SC_OK) //Ожидание статус-кода 200
                 .and()
                 .body("orders", notNullValue()); //Ожидание того, что поле orders в теле ответе не будет пустым
     }
